update SfvipUserProxy
Former-commit-id: 228c023fadbcfbb99f6dee309104c7454e564b34

+++ src/sfvip/update/__init__.py
@@ -4,48 +4,42 @@ import sys
 import tempfile
 from pathlib import Path
 from typing import Optional
-from urllib.error import ContentTooShortError, HTTPError, URLError
 
+from ..ui.progress import ProgressWindow
 from .cpu import Cpu
+from .download import download_and_unpack, download_in_thread
 from .libmpv import LibmpvDll
-from .progress import Progress
 
 logger = logging.getLogger(__name__)
 
 
-class PlayerUpdate:
+class _PlayerUpdate:
     bitness = "x64" if Cpu.is64 else "x86"
     _url = f"https://raw.githubusercontent.com/K4L4Uz/SFVIP-Player/master/Update_{bitness}.zip"
 
-    @staticmethod
-    def download(player_exe: Path, progress: Progress) -> bool:
+    def __init__(self, player_exe: Path) -> None:
+        self._player_exe = player_exe
+
+    def download(self, progress: ProgressWindow) -> bool:
         with tempfile.TemporaryDirectory() as temp_dir:
             archive = Path(temp_dir) / "player update"
-            progress.download_and_unpack(PlayerUpdate._url, archive, player_exe.parent)
-            if player_exe.exists():
-                logger.info("player exe found")
-                return True
+            if download_and_unpack(_PlayerUpdate._url, archive, self._player_exe.parent, progress):
+                if self._player_exe.exists():
+                    logger.info("player exe found")
+                    return True
         return False
 
 
 def download_player(player_name: str) -> Optional[str]:
-    def run() -> bool:
-        if PlayerUpdate.download(player_exe, progress):
-            progress.msg("Check latest libmpv")
-            libmpv_dll = LibmpvDll(player_exe)
-            if libmpv := libmpv_dll.check():
-                return libmpv_dll.download(libmpv, progress)
+    def download(progress: ProgressWindow) -> bool:
+        if _PlayerUpdate(player_exe).download(progress):
+            return LibmpvDll(player_exe).check_and_download(progress)
         return False
 
-    exceptions = OSError, URLError, HTTPError, ContentTooShortError, ValueError, shutil.ReadError
     exe_dir = Path(sys.argv[0]).parent
-    player_dir = exe_dir / f"{player_name.capitalize()} {PlayerUpdate.bitness}"
-    progress = Progress(f"Download {player_dir.name}", 400, *exceptions)
+    player_dir = exe_dir / f"{player_name.capitalize()} {_PlayerUpdate.bitness}"
     player_exe = player_dir / f"{player_name}.exe"
-    try:
-        if progress.run_in_thread(run, *exceptions, mainloop=True):
-            return str(player_exe)
-    except exceptions as err:
-        logger.warning("player download exception %s", err)
+    if download_in_thread(f"Download {player_dir.name}", download, create_mainloop=True):
+        return str(player_exe)
     shutil.rmtree(player_dir, ignore_errors=True)
     logger.warning("player download failed")
